pyproject.toml
[tool.poetry]
name = "parking_api"
version = "0.1.0"
description = "A test to coding a simple api of a parking"
authors = ["Gustavo Soares <gustavo.soares.cdc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
humanize = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
